requirements.txt
asyncio
cowpy
discord.py
python_dateutil
pytz
pyutil
requests
